# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: CIBUS API Sistema de Gestión de Servicios Gastronómicos
  version: 1.0.0
  description: >
    Documentación de la API REST para el sistema de gestión de servicios gastronómicos CIBUS.
servers:
  - url: http://localhost:8080/api
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get products
      description: Returns a paginated list of products.
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: categoryName
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'

    post:
      tags:
        - Products
      summary: Create product
      description: Creates a new product.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'

    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDTO'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDTO'

    delete:
      tags:
        - Products
      summary: Delete product
      parameters:
        - $ref: '#/components/parameters/productId'
      responses:
        '204':
          description: Product deleted successfully
...
  /product-categories:
    get:
      tags:
        - Product Categories
      summary: Get all product categories
      description: Returns a list of all product categories.
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryResponseDTO'

    post:
      tags:
        - Product Categories
      summary: Create product category
      description: Creates a new product category.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryRequestDTO'
      responses:
        '201':
          description: Product category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryResponseDTO'
  /product-categories/{productCategoryId}:
    get:
      tags:
        - Product Categories
      summary: Get product category by ID
      parameters:
        - $ref: '#/components/parameters/productCategoryId'
      responses:
        '200':
          description: Product category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryResponseDTO'

    put:
      tags:
        - Product Categories
      summary: Update product category
      parameters:
        - $ref: '#/components/parameters/productCategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCategoryRequestDTO'
      responses:
        '200':
          description: Product category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryResponseDTO'

    delete:
      tags:
        - Product Categories
      summary: Delete product category
      parameters:
        - $ref: '#/components/parameters/productCategoryId'
      responses:
        '204':
          description: Product category deleted successfully
  /product-categories/name/{name}:
    get:
      tags:
        - Product Categories
      summary: Get product category by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryResponseDTO'


  # ── BRANCH PRODUCTS ────────────────────────────────────────────────────────────
...
  /branches/{branchId}/products/{productId}:
    get:
      tags:
        - Branch Products
      summary: Get branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      responses:
        '200':
          description: Product found

    put:
      tags:
        - Branch Products
      summary: Update branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchProductUpdateDTO'
      responses:
        '200':
          description: Product updated successfully

    delete:
      tags:
        - Branch Products
      summary: Delete branch product
      parameters:
        - $ref: '#/components/parameters/branchId'
        - $ref: '#/components/parameters/productId'
      responses:
        '204':
          description: Product removed successfully

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /payment-methods:
    get:
      tags:
        - Payment Methods
      summary: Get all payment methods
      responses:
        '200':
          description: Payment methods retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponseDTO'

    post:
      tags:
        - Payment Methods
      summary: Create payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodRequestDTO'
      responses:
        '201':
          description: Payment method created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponseDTO'
  /payment-methods/{paymentMethodId}:
    get:
      tags:
        - Payment Methods
      summary: Get payment method by ID
      parameters:
        - $ref: '#/components/parameters/paymentMethodId'
      responses:
        '200':
          description: Payment method found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponseDTO'

    put:
      tags:
        - Payment Methods
      summary: Update payment method
      parameters:
        - $ref: '#/components/parameters/paymentMethodId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMethodRequestDTO'
      responses:
        '200':
          description: Payment method updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponseDTO'

    delete:
      tags:
        - Payment Methods
      summary: Delete payment method
      parameters:
        - $ref: '#/components/parameters/paymentMethodId'
      responses:
        '204':
          description: Payment method deleted successfully
  /payment-methods/name/{name}:
    get:
      tags:
        - Payment Methods
      summary: Search payment methods by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment methods found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponseDTO'             

# ── STATISTICS ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    branchId:
      name: branchId
      in: path
      required: true
      description: ID de la sucursal
      schema:
        type: integer
        example: 1
...
    productId:
      name: productId
      in: path
      required: true
      description: ID del producto
      schema:
        type: integer
        example: 10
    productCategoryId:
      name: productCategoryId
      in: path
      required: true
      description: Product category ID
      schema:
        type: integer
        format: int64
        example: 1    
...
    paymentMethodId:
      name: paymentMethodId
      in: path
      required: true
      description: Payment method ID
      schema:
        type: integer
        format: int64
        example: 1    
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    
  schemas:
...
    ProductRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "Milanesa Napolitana"
        description:
          type: string
          maxLength: 255
          example: "Breaded beef with tomato sauce and mozzarella cheese"
        categoryId:
          type: integer
          format: int64
          example: 2
      required:
        - name
        - categoryId
    ProductResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Milanesa Napolitana"
        description:
          type: string
          example: "Breaded beef with tomato sauce and mozzarella cheese"
        categoryName:
          type: string
          example: "Main Course"

  # ── PRODUCT CATEGORY ──────────────────────────────────────────────────────────────
    ProductCategoryRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "Main Course"
      required:
        - name
    ProductCategoryResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Main Course"

  # ── BRANCH PRODUCTS ──────────────────────────────────────────────────────────────
...
    BranchProductUpdateDTO:
      type: object
      properties:
        price:
          type: number
          format: decimal
          example: 2800.00
        available:
          type: boolean
          example: true
...
    PaymentMethodRequestDTO:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "Credit Card"
      required:
        - name
    PaymentMethodResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Credit Card"

    # ── STATISTICS ──────────────────────────────────────────────────────────────── 
...